Combine into a single mcp server
